Extracted Set class
Set was extracted from Resolution to handle all functionality once
structure had been formed.
