+++ .gitignore
@@ -61,3 +61,4 @@ Temporary Items
 /LogicParser.class
 /Main.class
 /Resolution.class
+/Clause.class
