+++ resolution/.gitignore
@@ -13,3 +13,4 @@
 /LogicParser$AntlrParser.class
 /LogicParser.class
 /DrawableTree.class
+/Set.class
